переправил в соответствии с требованиями

+++ src/main/java/com/javamentor/qa/platform/webapp/controllers/rest/ResourceAnswerController.java
@@ -1,12 +1,17 @@
 package com.javamentor.qa.platform.webapp.controllers.rest;
 
+import com.javamentor.qa.platform.models.entity.question.answer.Answer;
 import com.javamentor.qa.platform.service.abstracts.model.AnswerService;
 import io.swagger.annotations.*;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
+
+import java.util.Optional;
 
 @RestController
 @RequiredArgsConstructor
@@ -19,7 +24,11 @@ public class ResourceAnswerController {
     @ApiResponses(value =
     @ApiResponse(code = 200, message = "Успешное удаление"))
     @DeleteMapping("/{answerId}")
-    public void deleteAnswerById(@ApiParam("Id ответа") @PathVariable Long answerId) {
+    public ResponseEntity<String> deleteAnswerById(@ApiParam("Id ответа") @PathVariable Long answerId){
+        if (!answerService.existsById(answerId)) {
+            return ResponseEntity.badRequest().body("Answer with this ID was not found");
+        }
         answerService.deleteById(answerId);
+        return ResponseEntity.status(HttpStatus.OK).body("Answer successfully deleted");
     }
 }
